+++ tests/test_task.py
@@ -785,7 +785,6 @@ class TaskGraphTests(unittest.TestCase):
         logger_name = 'test.task.queuelogger'
         log_message = 'This is coming from another process'
         logger = logging.getLogger(logger_name)
-        logger.setLevel(logging.DEBUG)
         file_log_path = os.path.join(
             self.workspace_dir, 'test_multiprocessed_logging.log')
         file_handler = logging.FileHandler(file_log_path)
@@ -794,13 +793,24 @@ class TaskGraphTests(unittest.TestCase):
         logger.addHandler(file_handler)
 
         task_graph = taskgraph.TaskGraph(self.workspace_dir, 1)
+        # this first message shouldn't print because logger level is a warning
+        # and _log_from_another_process is info level
+        logger.setLevel(logging.WARNING)
+        log_task = task_graph.add_task(
+            func=_log_from_another_process,
+            args=(logger_name, 'YOU WILL NOT SEE THIS'))
+        log_task.join()
+
+        # second should log because logger level is set to info
+        logger.setLevel(logging.INFO)
         log_task = task_graph.add_task(
             func=_log_from_another_process,
             args=(logger_name, log_message))
         log_task.join()
-        file_handler.flush()
+
         task_graph.close()
         task_graph.join()
+        file_handler.flush()
         file_handler.close()
 
         @retrying.retry(wait_exponential_multiplier=100,
@@ -809,7 +819,6 @@ class TaskGraphTests(unittest.TestCase):
         def get_name_and_message():
             with open(file_log_path, 'r') as log_file:
                 message = log_file.read().rstrip()
-            print(message)
             process_name, logged_message = re.match(
                 ':([^:]*):([^:]*):', message).groups()
             return process_name, logged_message