+++ HISTORY.rst
@@ -23,6 +23,8 @@ Unreleased Changes
   recomputation. Rather than add additional complexity around the limitations
   of this feature it is being removed to guide a design toward a standalone
   cache library if needed.
+* Fixed issue where ``TaskGraph`` ignored the parent logger level and would
+  log messages with a lower priority than its effective level.
 
 0.10.3 (2021-01-29)
 -------------------